+++ src/azure-cli/azure/cli/command_modules/monitor/tests/latest/test_monitor_log_analytics_workspace.py
@@ -54,7 +54,7 @@ class TestLogProfileScenarios(ScenarioTest):
             self.check("@[?name=='AzureSecurityOfThings'].enabled", '[False]')
         ])
 
-        self.cmd("monitor log-analytics workspace delete -g {rg} -n {name}")
+        self.cmd("monitor log-analytics workspace delete -g {rg} -n {name} -y")
 
     @record_only()
     def test_monitor_log_analytics_workspace_linked_service_common_scenario(self):
@@ -110,10 +110,14 @@ class TestLogProfileScenarios(ScenarioTest):
 
     @ResourceGroupPreparer(name_prefix='cli_test_monitor_workspace_linked_storage', location='eastus')
     @AllowLargeResponse()
-    @StorageAccountPreparer(name_prefix='saws1', kind='StorageV2', sku='Standard_LRS', parameter_name='account_1', location='eastus')
-    @StorageAccountPreparer(name_prefix='saws2', kind='StorageV2', sku='Standard_LRS', parameter_name='account_2', location='eastus')
-    @StorageAccountPreparer(name_prefix='saws3', kind='StorageV2', sku='Standard_LRS', parameter_name='account_3', location='eastus')
-    @StorageAccountPreparer(name_prefix='saws4', kind='StorageV2', sku='Standard_LRS', parameter_name='account_4', location='eastus')
+    @StorageAccountPreparer(name_prefix='saws1', kind='StorageV2', sku='Standard_LRS', parameter_name='account_1',
+                            location='eastus')
+    @StorageAccountPreparer(name_prefix='saws2', kind='StorageV2', sku='Standard_LRS', parameter_name='account_2',
+                            location='eastus')
+    @StorageAccountPreparer(name_prefix='saws3', kind='StorageV2', sku='Standard_LRS', parameter_name='account_3',
+                            location='eastus')
+    @StorageAccountPreparer(name_prefix='saws4', kind='StorageV2', sku='Standard_LRS', parameter_name='account_4',
+                            location='eastus')
     def test_monitor_log_analytics_workspace_linked_storage(self, resource_group, account_1,
                                                             account_2, account_3, account_4):
         from msrestazure.tools import resource_id
@@ -240,3 +244,134 @@ class TestLogProfileScenarios(ScenarioTest):
             self.check('publicNetworkAccessForIngestion', 'Enabled'),
             self.check('publicNetworkAccessForQuery', 'Disabled')
         ])
+
+    @ResourceGroupPreparer(name_prefix='cli_test_monitor_workspace_recover', location='westus')
+    @AllowLargeResponse()
+    def test_monitor_log_analytics_workspace_recover(self, resource_group):
+        workspace_name = self.create_random_name('clitest', 20)
+        self.kwargs.update({
+            'name': workspace_name
+        })
+
+        self.cmd("monitor log-analytics workspace create -g {rg} -n {name} --quota 1 --level 100 --sku CapacityReservation", checks=[
+            self.check('provisioningState', 'Succeeded'),
+            self.check('retentionInDays', 30),
+            self.check('sku.name', 'capacityreservation'),
+            self.check('sku.capacityReservationLevel', 100),
+            self.check('workspaceCapping.dailyQuotaGb', 1.0)
+        ])
+
+        self.cmd("monitor log-analytics workspace update -g {rg} -n {name} --quota 2 --level 200", checks=[
+            self.check('provisioningState', 'Succeeded'),
+            self.check('sku.capacityReservationLevel', 200),
+            self.check('workspaceCapping.dailyQuotaGb', 2.0)
+        ])
+
+        self.cmd("monitor log-analytics workspace table list -g {rg} --workspace-name {name}", checks=[
+            # self.check('length(@)', 0)
+        ])
+
+        self.kwargs.update({
+            'table_name': 'Syslog'
+        })
+        self.cmd("monitor log-analytics workspace table update -g {rg} --workspace-name {name} -n {table_name} --retention-time 30 --debug", checks=[
+            self.check('retentionInDays', 30)
+        ])
+
+        self.cmd("monitor log-analytics workspace list-deleted-workspaces -g {rg}", checks=[
+            self.check('length(@)', 0)
+        ])
+
+        self.cmd("monitor log-analytics workspace delete -g {rg} -n {name} -y")
+
+        self.cmd("monitor log-analytics workspace list-deleted-workspaces -g {rg}", checks=[
+            self.check('length(@)', 1)
+        ])
+
+        self.cmd("monitor log-analytics workspace recover -g {rg} -n {name}", checks=[
+            self.check('provisioningState', 'Succeeded'),
+            self.check('retentionInDays', 30),
+            self.check('sku.name', 'capacityreservation'),
+            self.check('sku.capacityReservationLevel', 200),
+            self.check('workspaceCapping.dailyQuotaGb', 2.0)
+        ])
+
+        self.cmd("monitor log-analytics workspace show -g {rg} -n {name}", checks=[
+            self.check('provisioningState', 'Succeeded'),
+            self.check('retentionInDays', 30),
+            self.check('sku.name', 'capacityreservation'),
+            self.check('sku.capacityReservationLevel', 200),
+            self.check('workspaceCapping.dailyQuotaGb', 2.0)
+        ])
+
+        self.cmd("monitor log-analytics workspace list-deleted-workspaces -g {rg}", checks=[
+            self.check('length(@)', 0)
+        ])
+
+        self.cmd("monitor log-analytics workspace delete -g {rg} -n {name} --force -y")
+
+        with self.assertRaisesRegexp(SystemExit, '3'):
+            self.cmd('monitor log-analytics workspace show -g {rg} -n {name}')
+
+    @ResourceGroupPreparer(name_prefix='cli_test_monitor_workspace_saved_search', location='eastus')
+    def test_monitor_log_analytics_workspace_saved_search(self, resource_group):
+        self.kwargs.update({
+            'workspace_name': self.create_random_name('clitest', 20),
+            'saved_search_name': 'clitest',
+            'category': 'cli',
+            'category_2': 'cli2',
+            'query': "Heartbeat | getschema",
+            'query_2': "AzureActivity | summarize count() by bin(timestamp, 1h)",
+            'display_name': 'myclitest',
+            'display_name_2': 'myclitest2',
+            'function_alias': 'myfun',
+            'function_param': "a:string = value",
+            'function_alias_2': 'myfun2',
+            'function_param_2': "a2:string = value",
+            'rg': resource_group
+        })
+
+        self.cmd("monitor log-analytics workspace create -g {rg} -n {workspace_name} --tags clitest=myron")
+
+        self.cmd('monitor log-analytics workspace saved-search create -g {rg} --workspace-name {workspace_name} -n {saved_search_name} '
+                 '--category {category} --display-name {display_name} -q "{query}" --fa {function_alias} '
+                 '--fp "{function_param}" --tags a=b c=d',
+                 checks=[
+                     self.check('category', '{category}'),
+                     self.check('displayName', '{display_name}'),
+                     self.check('query', "{query}"),
+                     self.check('functionAlias', '{function_alias}'),
+                     self.check('functionParameters', '{function_param}'),
+                     self.check('length(tags)', 2)
+                 ])
+
+        self.cmd('monitor log-analytics workspace saved-search show -g {rg} --workspace-name {workspace_name} -n {saved_search_name}', checks=[
+            self.check('category', '{category}'),
+            self.check('displayName', '{display_name}'),
+            self.check('query', "Heartbeat | getschema"),
+            self.check('functionAlias', '{function_alias}'),
+            self.check('functionParameters', '{function_param}'),
+            self.check('length(tags)', 2)
+        ])
+        self.cmd('monitor log-analytics workspace saved-search list -g {rg} --workspace-name {workspace_name}', checks=[
+            self.check('length(@)', 1)
+        ])
+
+        self.cmd(
+            'monitor log-analytics workspace saved-search update -g {rg} --workspace-name {workspace_name} -n {saved_search_name} '
+            '--category {category_2} --display-name {display_name_2} -q "{query_2}" --fa {function_alias_2} '
+            '--fp "{function_param_2}" --tags a=c f=e',
+            checks=[
+                self.check('category', '{category_2}'),
+                self.check('displayName', '{display_name_2}'),
+                self.check('query', "{query_2}"),
+                self.check('functionAlias', '{function_alias_2}'),
+                self.check('functionParameters', '{function_param_2}'),
+                self.check('length(tags)', 2),
+                self.check('tags[0].value', 'c'),
+                self.check('tags[1].value', 'e')
+            ])
+
+        self.cmd('monitor log-analytics workspace saved-search delete -g {rg} --workspace-name {workspace_name} -n {saved_search_name} -y')
+        with self.assertRaisesRegexp(SystemExit, '3'):
+            self.cmd('monitor log-analytics workspace saved-search show -g {rg} --workspace-name {workspace_name} -n {saved_search_name}')