+++ src/azure-cli/azure/cli/command_modules/resource/tests/latest/test_resource.py
@@ -55,6 +55,32 @@ class ResourceGroupScenarioTest(ScenarioTest):
 
         self.assertEqual('"1.0.0.0"\n', result.output)
 
+    @ResourceGroupPreparer(name_prefix='cli_test_rg_scenario')
+    def test_resource_group_export_skip_all_params(self, resource_group):
+
+        self.kwargs.update({
+            'vnet': 'vnet1'
+        })
+
+        self.cmd('network vnet create -g {rg} -n {vnet}')
+        self.kwargs['vnet_id'] = self.cmd('network vnet show -g {rg} -n {vnet}').get_output_in_json()['id']
+        result = self.cmd('group export --name {rg} --resource-ids "{vnet_id}" --skip-all-params --query "parameters"')
+
+        self.assertEqual('{}\n', result.output)
+
+    @ResourceGroupPreparer(name_prefix='cli_test_rg_scenario')
+    def test_resource_group_export_skip_resource_name_params(self, resource_group):
+
+        self.kwargs.update({
+            'vnet': 'vnet1'
+        })
+
+        self.cmd('network vnet create -g {rg} -n {vnet}')
+        self.kwargs['vnet_id'] = self.cmd('network vnet show -g {rg} -n {vnet}').get_output_in_json()['id']
+        result = self.cmd('group export --name {rg} --resource-ids "{vnet_id}" --skip-resource-name-params --query "parameters"')
+
+        self.assertEqual('{}\n', result.output)
+
 
 class ResourceGroupNoWaitScenarioTest(ScenarioTest):
 
@@ -1193,6 +1219,10 @@ class FeatureScenarioTest(ScenarioTest):
         # Once a feature goes GA , it will be removed from the feature list. Once that happens, use other ones to test
         self.cmd('feature show --namespace Microsoft.Network -n AllowLBPreview')
 
+    @AllowLargeResponse(8192)
+    def test_feature_unregister(self):
+        self.cmd('feature unregister --namespace Microsoft.Network --name AllowLBPreview', checks=self.check('properties.state', 'Unregistered'))
+
 
 class PolicyScenarioTest(ScenarioTest):
 
@@ -2259,14 +2289,26 @@ class ResourceGroupLocalContextScenarioTest(LocalContextScenarioTest):
 
     def test_resource_group_local_context(self):
         self.kwargs.update({
-            'group': 'test_local_context_group',
+            'group1': 'test_local_context_group_1',
+            'group2': 'test_local_context_group_2',
             'location': 'eastasia'
         })
-        self.cmd('group create -n {group} -l {location}', checks=[self.check('name', self.kwargs['group'])])
-        self.cmd('group show', checks=[
-            self.check('name', self.kwargs['group']),
+        self.cmd('group create -n {group1} -l {location}', checks=[
+            self.check('name', self.kwargs['group1']),
             self.check('location', self.kwargs['location'])
         ])
+        self.cmd('group show', checks=[
+            self.check('name', self.kwargs['group1']),
+            self.check('location', self.kwargs['location'])
+        ])
+        with self.assertRaisesRegexp(SystemExit, '2'):
+            self.cmd('group delete')
+        self.cmd('group delete -n {group1} -y')
+        self.cmd('group create -n {group2}', checks=[
+            self.check('name', self.kwargs['group2']),
+            self.check('location', self.kwargs['location'])
+        ])
+        self.cmd('group delete -n {group2} -y')
 
 
 if __name__ == '__main__':