+++ src/azure-cli/azure/cli/command_modules/security/commands.py
@@ -16,12 +16,33 @@ from ._client_factory import (cf_security_tasks,
                               cf_security_pricings,
                               cf_security_topology,
                               cf_security_workspace_settings,
-                              cf_security_advanced_threat_protection)
+                              cf_security_advanced_threat_protection,
+                              cf_security_assessment,
+                              cf_security_assessment_metadata,
+                              cf_security_sub_assessment,
+                              cf_security_regulatory_compliance_standards,
+                              cf_security_regulatory_compliance_control,
+                              cf_security_regulatory_compliance_assessment)
 
 
 # pylint: disable=line-too-long
 # pylint: disable=too-many-statements
 def load_command_table(self, _):
+
+    security_regulatory_compliance_standards_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.security.operations#RegulatoryComplianceStandardsOperations.{}',
+        client_factory=cf_security_regulatory_compliance_standards
+    )
+
+    security_regulatory_compliance_controls_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.security.operations#RegulatoryComplianceControlsOperations.{}',
+        client_factory=cf_security_regulatory_compliance_control
+    )
+
+    security_regulatory_compliance_assessment_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.security.operations#RegulatoryComplianceAssessmentsOperations.{}',
+        client_factory=cf_security_regulatory_compliance_assessment
+    )
 
     security_tasks_sdk = CliCommandType(
         operations_tmpl='azure.mgmt.security.operations#TasksOperations.{}',
@@ -99,6 +120,39 @@ def load_command_table(self, _):
         operations_tmpl='azure.mgmt.security.operations#AdvancedThreatProtectionOperations.{}',
         client_factory=cf_security_advanced_threat_protection
     )
+
+    security_assessment_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.security.operations#AssessmentsOperations.{}',
+        client_factory=cf_security_assessment
+    )
+
+    security_assessment_metadata_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.security.operations#AssessmentMetadataOperations.{}',
+        client_factory=cf_security_assessment_metadata
+    )
+
+    security_sub_assessment_sdk = CliCommandType(
+        operations_tmpl='azure.mgmt.security.operations#SubAssessmentsOperations.{}',
+        client_factory=cf_security_sub_assessment
+    )
+
+    with self.command_group('security regulatory-compliance-standards',
+                            security_regulatory_compliance_standards_sdk,
+                            client_factory=cf_security_regulatory_compliance_standards) as g:
+        g.custom_command('list', 'list_regulatory_compliance_standards')
+        g.custom_command('show', 'get_regulatory_compliance_standard')
+
+    with self.command_group('security regulatory-compliance-controls',
+                            security_regulatory_compliance_controls_sdk,
+                            client_factory=cf_security_regulatory_compliance_control) as g:
+        g.custom_command('list', 'list_regulatory_compliance_controls')
+        g.custom_command('show', 'get_regulatory_compliance_control')
+
+    with self.command_group('security regulatory-compliance-assessments',
+                            security_regulatory_compliance_assessment_sdk,
+                            client_factory=cf_security_regulatory_compliance_assessment) as g:
+        g.custom_command('list', 'list_regulatory_compliance_assessments')
+        g.custom_command('show', 'get_regulatory_compliance_assessment')
 
     with self.command_group('security task',
                             security_tasks_sdk,
@@ -184,6 +238,28 @@ def load_command_table(self, _):
         g.custom_command('show', 'get_security_workspace_setting')
         g.custom_command('create', 'create_security_workspace_setting')
         g.custom_command('delete', 'delete_security_workspace_setting')
+
+    with self.command_group('security assessment',
+                            security_assessment_sdk,
+                            client_factory=cf_security_assessment) as g:
+        g.custom_command('list', 'list_security_assessments')
+        g.custom_command('show', 'get_security_assessment')
+        g.custom_command('create', 'create_security_assessment')
+        g.custom_command('delete', 'delete_security_assessment')
+
+    with self.command_group('security assessment-metadata',
+                            security_assessment_metadata_sdk,
+                            client_factory=cf_security_assessment_metadata) as g:
+        g.custom_command('list', 'list_security_assessment_metadata')
+        g.custom_command('show', 'get_security_assessment_metadata')
+        g.custom_command('create', 'create_security_assessment_metadata')
+        g.custom_command('delete', 'delete_security_assessment_metadata')
+
+    with self.command_group('security sub-assessment',
+                            security_sub_assessment_sdk,
+                            client_factory=cf_security_sub_assessment) as g:
+        g.custom_command('list', 'list_security_sub_assessments')
+        g.custom_command('show', 'get_security_sub_assessment')
 
     with self.command_group('security', is_preview=True):
         pass