+++ src/azure-cli/azure/cli/command_modules/sql/_util.py
@@ -178,3 +178,11 @@ def get_sql_managed_database_long_term_retention_policies_operations(cli_ctx, _)
 
 def get_sql_managed_database_long_term_retention_backups_operations(cli_ctx, _):
     return get_sql_management_client(cli_ctx).long_term_retention_managed_instance_backups
+
+
+def get_sql_database_long_term_retention_policies_operations(cli_ctx, _):
+    return get_sql_management_client(cli_ctx).backup_long_term_retention_policies
+
+
+def get_sql_database_long_term_retention_backups_operations(cli_ctx, _):
+    return get_sql_management_client(cli_ctx).long_term_retention_backups