+++ src/azure-cli/azure/cli/command_modules/storage/oauth_token_util.py
@@ -25,19 +25,20 @@ class TokenUpdater(object):
         from azure.cli.core._profile import Profile
         from datetime import datetime
         # should give back token that is valid for at least 5 mins
-        token = Profile(cli_ctx=self.cli_ctx).get_raw_token(resource="https://storage.azure.com")[0][2]
+        token = Profile(cli_ctx=self.cli_ctx).get_raw_token(
+            resource="https://storage.azure.com", subscription=self.cli_ctx.data['subscription_id'])[0][2]
         try:
             self.token_credential.token = token['accessToken']
             seconds_left = (datetime.strptime(token['expiresOn'], "%Y-%m-%d %H:%M:%S.%f") - datetime.now()).seconds
         except KeyError:  # needed to deal with differing unserialized MSI token payload
             self.token_credential.token = token['access_token']
             seconds_left = (datetime.fromtimestamp(int(token['expires_on'])) - datetime.now()).seconds
-        if seconds_left < 240:
-            # acquired token expires in less than 4 mins
-            raise Exception("Acquired a token expiring in less than 4 minutes")
+        if seconds_left < 180:
+            # acquired token expires in less than 3 mins
+            raise Exception("Acquired a token expiring in less than 3 minutes")
 
         with self.lock:
-            self.timer = threading.Timer(seconds_left - 240, self.timer_callback)
+            self.timer = threading.Timer(seconds_left - 180, self.timer_callback)
             self.timer.daemon = True
             self.timer.start()